requirements.txt
streamlit
pymongo
PyYAML
matplotlib
pandas
numpy
openpyxl
xlrd
scikit-learn
python-dotenv